+++ lib/src/result.dart
@@ -1,13 +1,10 @@
-import 'package:meta/meta.dart';
-
 class BraintreeDropInResult {
   const BraintreeDropInResult({
-    @required this.paymentMethodNonce,
-    @required this.deviceData,
+    required this.paymentMethodNonce,
+    required this.deviceData,
   });
 
   factory BraintreeDropInResult.fromJson(dynamic source) {
-    if (source == null) return null;
     return BraintreeDropInResult(
       paymentMethodNonce:
           BraintreePaymentMethodNonce.fromJson(source['paymentMethodNonce']),
@@ -19,24 +16,25 @@ class BraintreeDropInResult {
   final BraintreePaymentMethodNonce paymentMethodNonce;
 
   /// String of device data. `null`, if `collectDeviceData` was set to false.
-  final String deviceData;
+  final String? deviceData;
 }
 
 class BraintreePaymentMethodNonce {
   const BraintreePaymentMethodNonce({
-    @required this.nonce,
-    @required this.typeLabel,
-    @required this.description,
-    @required this.isDefault,
+    required this.nonce,
+    required this.typeLabel,
+    required this.description,
+    required this.isDefault,
+    this.paypalPayerId,
   });
 
   factory BraintreePaymentMethodNonce.fromJson(dynamic source) {
-    if (source == null) return null;
     return BraintreePaymentMethodNonce(
       nonce: source['nonce'],
       typeLabel: source['typeLabel'],
       description: source['description'],
       isDefault: source['isDefault'],
+      paypalPayerId: source['paypalPayerId'],
     );
   }
 
@@ -52,4 +50,7 @@ class BraintreePaymentMethodNonce {
 
   /// True if this payment method is the default for the current customer, false otherwise.
   final bool isDefault;
+
+  /// PayPal payer id if requesting for paypal nonce
+  final String? paypalPayerId;
 }