+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## 2.2.0
+
+* Add PayPal Payer ID to result object (thank you to [nabinadhikari](https://github.com/nabinadhikari)!)
+
+## 2.1.0
+* Fix PayPal vault flow not working on iOS (thank you to [andrea689](https://github.com/andrea689)!)
+* Add support for Apple Pay's `PKPaymentSummaryItem` (thank you to [bkovac](https://github.com/bkovac)!)
+
+## 2.0.0+1
+
+* Fix new build issue on iOS (thank you to [JideGuru](https://github.com/JideGuru)!)
+
+## 2.0.0
+
+* Upgrade several dependencies to fix build issues on Android and iOS (thank you to [bennibau](https://github.com/bennibau), [reverie-ss](https://github.com/reverie-ss), and [andesappal](https://github.com/andesappal)!)
+
+## 2.0.0-nullsafety.0
+
+* Add null-safety support
+
+## 1.2.1
+
+* Fix Android build (hopefully)
+
+## 1.2.0
+
+* Add option for CreditCard CVV number (thank you to [ilicmilan](https://github.com/ilicmilan)!)
+
+## 1.1.0+1
+
+* Fix vulnerability in Braintree plugin (Play Store issue)
+
+## 1.1.0
+
+* Add ApplePay support for the drop-in UI (thank you again to HareshGediya!)
+
+## 1.0.3
+
+* Fix `cardEnabled = false` not working on iOS drop-in (thank you to HareshGediya)
+
 ## 1.0.2+2
 
 * Hotfix 3D Secure (thank you to enzobonggio!)